парсинг веб/requirements.txt
selenium
webdriver-manager
python-telegram-bot
undetected-chromedriver
